More renaming to fix test failures.
