+++ pyrostest/__init__.py
@@ -1,5 +1,8 @@
 """Collection of utilities for testing ros nodes less painfully.
 """
 
+import pyrostest.rostest_utils
+
 from pyrostest.launch_tools import with_launch_file, launch_node
-from pyrostest.pyrostest import RosTest
+from pyrostest.ros_test import RosTest
+
